@@ -2,7 +2,7 @@
   "name": "marked",
   "description": "A markdown parser built for speed",
   "author": "Christopher Jeffrey",
-  "version": "9.1.2",
+  "version": "9.1.5",
   "type": "module",
   "main": "./lib/marked.cjs",
   "module": "./lib/marked.esm.js",
@@ -20,9 +20,14 @@
   ],
   "exports": {
     ".": {
-      "types": "./lib/marked.d.ts",
-      "import": "./lib/marked.esm.js",
-      "default": "./lib/marked.cjs"
+      "import": {
+        "types": "./lib/marked.d.ts",
+        "default": "./lib/marked.esm.js"
+      },
+      "default": {
+        "types": "./lib/marked.d.cts",
+        "default": "./lib/marked.cjs"
+      }
     },
     "./bin/marked": "./bin/marked.js",
     "./marked.min.js": "./marked.min.js",
@@ -45,6 +50,7 @@
     "html"
   ],
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.12.2",
     "@markedjs/html-differ": "^4.0.2",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^11.1.5",
@@ -53,25 +59,26 @@
     "@semantic-release/github": "^9.2.1",
     "@semantic-release/npm": "^11.0.0",
     "@semantic-release/release-notes-generator": "^12.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
+    "@typescript-eslint/eslint-plugin": "^6.9.0",
+    "@typescript-eslint/parser": "^6.9.0",
     "cheerio": "^1.0.0-rc.12",
     "commonmark": "0.30.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.50.0",
+    "dts-bundle-generator": "^8.0.1",
+    "eslint": "^8.52.0",
     "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-n": "^16.1.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-n": "^16.2.0",
     "eslint-plugin-promise": "^6.1.1",
     "front-matter": "^4.0.2",
     "highlight.js": "^11.9.0",
     "jasmine": "^5.1.0",
     "markdown-it": "13.0.2",
     "marked-highlight": "^2.0.6",
-    "marked-man": "^1.3.4",
+    "marked-man": "^2.0.0",
     "node-fetch": "^3.3.2",
     "recheck": "^4.4.5",
-    "rollup": "^4.0.2",
+    "rollup": "^4.1.5",
     "semantic-release": "^22.0.5",
     "titleize": "^4.0.0",
     "ts-expect": "^1.3.0",
@@ -85,7 +92,7 @@
     "test:specs": "npm test -- test/specs/**/*-spec.js",
     "test:lint": "eslint .",
     "test:redos": "node test/recheck.js > vuln.js",
-    "test:types": "tsc --project tsconfig-type-test.json",
+    "test:types": "tsc --project tsconfig-type-test.json && attw -P --exclude-entrypoints ./bin/marked ./marked.min.js",
     "test:umd": "node test/umd-test.js",
     "test:update": "node test/update-specs.js",
     "rules": "node test/rules.js",
@@ -94,7 +101,7 @@
     "build:reset": "git checkout upstream/master lib/marked.cjs lib/marked.umd.js lib/marked.esm.js marked.min.js",
     "build": "npm run rollup && npm run build:types && npm run build:man",
     "build:docs": "npm run build && node docs/build.js",
-    "build:types": "tsc --project tsconfig-types.json",
+    "build:types": "tsc && dts-bundle-generator --project tsconfig.json -o lib/marked.d.ts src/marked.ts && dts-bundle-generator --project tsconfig.json -o lib/marked.d.cts src/marked.ts",
     "build:man": "marked-man man/marked.1.md > man/marked.1",
     "rollup": "rollup -c rollup.config.js"
   },
